+++ 1.12/src/main/java/panszelescik/moreplates/helpers/Strings.java
@@ -18,6 +18,7 @@ public class Strings {
 	public static String GOLD = "Gold";
 	public static String STEEL = "Steel";
 	public static String REDSTONE = "Redstone";
+	public static String GLOWSTONE = "Glowstone";
 	public static String COAL = "Coal";
 	public static String DIAMOND = "Diamond";
 	public static String EMERALD = "Emerald";
@@ -45,26 +46,28 @@ public class Strings {
 	public static String INFO_8 = " for: ";
 	public static String INFO_9 = " to: ";
 	
+	private static String COMPACTOR_NAME = "Compactor";
 	private static String ELVEN_TRADE_NAME = "Elven Trade";
 	private static String EMPOWERER_NAME = "Empowerer";
 	private static String IC2_CUTTING_NAME = "Block Cutting Machine";
 	private static String IC2_FORMER_NAME = "Metal Former";
 	private static String IE_MACHINE_NAME = "Metal Press";
+	private static String INDUCTION_SMELTER_NAME = "Induction Smelter";
 	private static String MANA_INFUSION_NAME = "Mana Infusion";
 	private static String RECONSTRUCTOR_NAME = "Atomic Reconstructor";
-	private static String TE_MACHINE_NAME = "Compactor";
 	private static String TECHREBORN_MACHINE_NAME = "Compressor";
 	
 	private static String INFO_END = INFO_2 + INFO_8;
 	
+	public static String INFO_COMPACTOR = INFO_1 + COMPACTOR_NAME + INFO_END;
 	public static String INFO_CUTTING_IC2 = INFO_1 + IC2_CUTTING_NAME + INFO_END;
 	public static String INFO_ELVEN_TRADE = INFO_1 + ELVEN_TRADE_NAME + INFO_END;
 	public static String INFO_EMPOWERER = INFO_1 + EMPOWERER_NAME + INFO_END;
 	public static String INFO_FORMER_IC2 = INFO_1 + IC2_FORMER_NAME + INFO_END;
 	public static String INFO_IE = INFO_1 + IE_MACHINE_NAME + INFO_END;
+	public static String INFO_INDUCTION_SMELTER = INFO_1 + INDUCTION_SMELTER_NAME + INFO_END;
 	public static String INFO_MANA_INFUSION = INFO_1 + MANA_INFUSION_NAME + INFO_END;
 	public static String INFO_RECONSTRUCTOR = INFO_1 + RECONSTRUCTOR_NAME + INFO_END;
-	public static String INFO_TE = INFO_1 + TE_MACHINE_NAME + INFO_END;
 	public static String INFO_TECHREBORN = INFO_1 + TECHREBORN_MACHINE_NAME + INFO_END;
 	public static String INFO_REG_ITEM = INFO_4 + INFO_5;
 	public static String INFO_REG_RENDER = INFO_4 + INFO_6;